+++ martin/src/config/file/file_config.rs
@@ -10,68 +10,8 @@ use martin_core::tiles::BoxedSource;
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-use crate::config::file::ConfigFileError::{InvalidFilePath, InvalidSourceUrl, IoError};
+use crate::config::file::{ConfigFileError, ConfigFileResult};
 use crate::{MartinError, MartinResult};
-
-pub type ConfigFileResult<T> = Result<T, ConfigFileError>;
-
-#[derive(thiserror::Error, Debug)]
-pub enum ConfigFileError {
-    #[error("IO error {0}: {1}")]
-    IoError(#[source] std::io::Error, PathBuf),
-
-    #[error("Unable to load config file {1}: {0}")]
-    ConfigLoadError(#[source] std::io::Error, PathBuf),
-
-    #[error("Unable to parse config file {1}: {0}")]
-    ConfigParseError(#[source] subst::yaml::Error, PathBuf),
-
-    #[error("Unable to write config file {1}: {0}")]
-    ConfigWriteError(#[source] std::io::Error, PathBuf),
-
-    #[error(
-        "No tile sources found. Set sources by giving a database connection string on command line, env variable, or a config file."
-    )]
-    NoSources,
-    #[error("Source path is not a file: {0}")]
-    InvalidFilePath(PathBuf),
-
-    #[error("Error {0} while parsing URL {1}")]
-    InvalidSourceUrl(#[source] url::ParseError, String),
-
-    #[error("Could not parse source path {0} as a URL")]
-    PathNotConvertibleToUrl(PathBuf),
-
-    #[error("Source {0} uses bad file {1}")]
-    InvalidSourceFilePath(String, PathBuf),
-
-    #[error("At least one 'origin' must be specified in the 'cors' configuration")]
-    CorsNoOriginsConfigured,
-
-    #[cfg(feature = "styles")]
-    #[error("Walk directory error {0}: {1}")]
-    DirectoryWalking(#[source] walkdir::Error, PathBuf),
-
-    #[cfg(feature = "postgres")]
-    #[error("The postgres pool_size must be greater than or equal to 1")]
-    PostgresPoolSizeInvalid,
-
-    #[cfg(feature = "postgres")]
-    #[error("A postgres connection string must be provided")]
-    PostgresConnectionStringMissing,
-
-    #[cfg(feature = "postgres")]
-    #[error("Failed to create postgres pool: {0}")]
-    PostgresPoolCreationFailed(#[source] martin_core::tiles::postgres::PostgresError),
-
-    #[cfg(feature = "fonts")]
-    #[error("Failed to load fonts from {1}: {0}")]
-    FontResolutionFailed(#[source] martin_core::fonts::FontError, PathBuf),
-
-    #[cfg(feature = "pmtiles")]
-    #[error("Failed to parse object store URL of {1}: {0}")]
-    ObjectStoreUrlParsing(object_store::Error, String),
-}
 
 /// Lifecycle hooks for configuring the application
 ///
@@ -321,7 +261,8 @@ impl FileConfigSrc {
             return Ok(path.clone());
         }
 
-        path.canonicalize().map_err(|e| IoError(e, path.clone()))
+        path.canonicalize()
+            .map_err(|e| ConfigFileError::IoError(e, path.clone()))
     }
 }
 
@@ -417,12 +358,14 @@ async fn resolve_int<T: TileSourceConfiguration>(
             } else if path.is_file() {
                 vec![path]
             } else {
-                return Err(MartinError::from(InvalidFilePath(
+                return Err(MartinError::from(ConfigFileError::InvalidFilePath(
                     path.canonicalize().unwrap_or(path),
                 )));
             };
             for path in dir_files {
-                let can = path.canonicalize().map_err(|e| IoError(e, path.clone()))?;
+                let can = path
+                    .canonicalize()
+                    .map_err(|e| ConfigFileError::IoError(e, path.clone()))?;
                 if files.contains(&can) {
                     if !is_dir {
                         warn!("Ignoring duplicate MBTiles path: {}", can.display());
@@ -458,7 +401,7 @@ fn collect_files_with_extension(
 ) -> Result<Vec<PathBuf>, ConfigFileError> {
     Ok(base_path
         .read_dir()
-        .map_err(|e| IoError(e, base_path.to_path_buf()))?
+        .map_err(|e| ConfigFileError::IoError(e, base_path.to_path_buf()))?
         .filter_map(Result::ok)
         .filter(|f| {
             f.path()
@@ -498,7 +441,7 @@ fn parse_url(is_enabled: bool, path: &Path) -> Result<Option<Url>, ConfigFileErr
     ];
     path.to_str()
         .filter(|v| url_schemes.iter().any(|scheme| v.starts_with(scheme)))
-        .map(|v| Url::parse(v).map_err(|e| InvalidSourceUrl(e, v.to_string())))
+        .map(|v| Url::parse(v).map_err(|e| ConfigFileError::InvalidSourceUrl(e, v.to_string())))
         .transpose()
 }
