+++ martin/src/config/file/resources/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "fonts")]
+pub mod fonts;
+#[cfg(feature = "sprites")]
+pub mod sprites;
+#[cfg(feature = "styles")]
+pub mod styles;
