+++ martin/src/config/file/tiles/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "unstable-cog")]
+pub mod cog;
+#[cfg(feature = "mbtiles")]
+pub mod mbtiles;
+#[cfg(feature = "pmtiles")]
+pub mod pmtiles;
+#[cfg(feature = "postgres")]
+pub mod postgres;
