+++ martin/src/config/file/tiles/postgres/resolver/scripts/query_available_function.sql
